Clean up the feature file
We no longer need these two scenarios, here's why:

The first one is identical to the new scenario for proximity: The one
where the listener is within range.

The second one was used to "triangulate" the code to force us away from
having a hard-coded message. Now that we have a model that uses a
variable for the message, there's an insignificant chance of this
behaviour regressing, so we can safely remove the test.