Fix unit tests
We've had to put in a dirty hack here in the specs to get them to carry
on passing now that we have the credits functionality.

This hack is a classic symptom of test-last programming. If we'd TDD'd
this, we'd have noticed that the test was hard to write, listened to
that feedback, and reconsidered our design. Now, because we've already
written the messy code and this failing test is just an inconvenience,
our motivation is to just bash it into shape and move on.

So we get no design feedback (or we ignore it anyway) and we get a
messy hack in the test.