Move `deduct_credits` onto Person
We'll make it a private method soon. Our acceptance tests are all
passing, but the unit tests are still broken.

This is because Network is still taking responsibility for deducting
credits, even if it leaves the details to the Shouter. Let's get rid of
that altogether and move the call to deduct credits onto the shouter
too.