RED: Start to TDD the homepage
We've started by implementing the first action in the When step,
log_in_as which just opens up the homepage with the user's name in the
query string.

We're dropping down to unit tests for the web app now, so we get
fine-grained feedback about our progress.

Note that we've used constructor arguments for the Sinatra
Application sublcass to allow us to inject dependencies.

We've also separated the core - our inner hexagon - from the web app.