Push down strategy for retrieving heard messages
We ran a manual test and noticed that you can't actually see the
messages you're supposed to have heard, yet all the tests were passing.

It turned out this is because our Then step was still going direct to
the domain model.

We want to allow for different strategies. We moved the logic for
retrieving heard messages down into the World class.

Next we need to implement the strategy for the web world.